Add generic LoginFilterRenderer instead of HeaderRenderer.
