+++ src/main/App.test.js
@@ -5,7 +5,6 @@ import Adapter from 'enzyme-adapter-react-16';
 import Enzyme,{shallow} from 'enzyme';
 import App from './App';
 import {Footer,Header,Navigation,Page} from './layout';
-import HeaderRenderer from './renderers/HeaderRenderer';
 
 Enzyme.configure({adapter: new Adapter()});
 
@@ -30,7 +29,7 @@ describe('App contains',()=>{
   })
 
   test('Header component', ()=>{
-    const componentExists=wrapper.exists(HeaderRenderer);
+    const componentExists=wrapper.exists(Header);
     expect(componentExists).toBe(true);
   })
